NewGraphButton controls 2 sperate modals
