+++ cypress/e2e/alert.spec.cy.ts
@@ -33,7 +33,7 @@ beforeEach(() => {
   intelligentAIButton.click();
 
   cy.get("textarea").type("今天是个好日子");
-  const createGraphButton = cy.get(".modalContent>div>div>button");
+  const createGraphButton = cy.get(".newGraphMethodModal>div>div>button");
   createGraphButton.click();
 });
